requirements.txt
requests~=2.32.3
pandas~=2.2.3
matplotlib~=3.10.1
numpy~=2.2.4
pip-licenses~=5.0.0
plotly~=6.0.1
beautifulsoup4~=4.13.3
selenium~=4.32.0
pillow~=11.2.1
pytz~=2025.2